more

+++ packages/protocol/contracts/L1/libs/LibVerifying.sol
@@ -41,7 +41,7 @@ library LibVerifying {
         state.l2Hashes[0] = genesisBlockHash;
 
         require(proofVerifier != address(0), "L1:verifier");
-        state.lookups["proofVerifier"] = proofVerifier;
+        state.lookups["proofVerifier"] = uint256(uint160(proofVerifier));
 
         emit BlockVerified(0, genesisBlockHash);
         emit HeaderSynced(block.number, 0, genesisBlockHash);
